Rename command to Slice and update plug-in packaging
