+++ src/plugin/dev/Slice_cmd.py
@@ -11,11 +11,11 @@ _HERE = os.path.dirname(os.path.abspath(__file__))
 if _HERE not in sys.path:
     sys.path.insert(0, _HERE)
 
-from RhinoToSlicer.commands import send_to_prusa as _impl
+from RhinoToSlicer.commands import slice as _impl
 
 __commandname__ = _impl.COMMAND_NAME
 
 
 def RunCommand(is_interactive):
-    result = _impl.send_to_prusaslicer()
+    result = _impl.send_to_slicer()
     return result if isinstance(result, Result) else Result.Success
